Simply moved all module for screen switching to separate module
